refactor(vue): split OrderView into composable and dialog components
Extract order screen logic to useOrderView and product/customer dialogs so
OrderView.vue stays under the 1000 LOC target without changing UX.

Closes #339
